+++ TKMovesets2/GameInteractions/Extractors/Extractor_t7.hpp
@@ -47,7 +47,7 @@ private:
 
 public:
 	using Extractor::Extractor; // Inherit constructor too
-	ExtractionErrcode Extract(gameAddr playerAddress, uint8_t gameId, bool overwriteSameFilename, float& progress) override;
+	ExtractionErrcode Extract(gameAddr playerAddress, uint8_t gameId, bool overwriteSameFilename, uint8_t& progress) override;
 	bool CanExtract() override;
 
 	std::string GetPlayerCharacterName(gameAddr playerAddress) override;
